Shoot from anywhere(hopefully) -OC
If the Y button is press while the Right Bumper is pressed, then the robot should shoot at the angle it needs for that distance. Commented out because it might interfere with shoot from speaker = y, also because I'm not sure if null will work. Also made minor text changes.